Split files.go so find, member-write, and rename/delete logic each live in their own file.
